moved default printer from DOM.mjs to index.mjs
